+++ src/main/java/lists/PalindromeLinkedList.java
@@ -61,18 +61,18 @@ public class PalindromeLinkedList {
     }
 
     public static boolean isPalindrome(ListNode head) {
-        if(head==null) {
+        if (head == null) {
             return true;
         }
         List<Integer> cache = convertToList(head);
-        return checkValidPalindrom(cache, 0, cache.size()-1);
+        return checkValidPalindrom(cache, 0, cache.size() - 1);
     }
 
     private static boolean checkValidPalindrom(List<Integer> cache, int left, int right) {
-        if(left==right || (right-left==1 && cache.get(left)==cache.get(right))) {
+        if (left == right || (right - left == 1 && cache.get(left) == cache.get(right))) {
             return true;
-        } else if ((int)cache.get(left)==(int)cache.get(right)) {
-            if(right-left==1){
+        } else if (cache.get(left) == cache.get(right)) {
+            if (right - left == 1) {
                 return true;
             } else {
                 return checkValidPalindrom(cache, ++left, --right);
@@ -85,7 +85,7 @@ public class PalindromeLinkedList {
         List<Integer> cache = new ArrayList<Integer>();
         ListNode current = head;
 
-        while(current!=null) {
+        while (current != null) {
             cache.add(current.val);
             current = current.next;
         }
